+++ src/icalendar/cal.py
@@ -40,7 +40,7 @@ from icalendar.prop import (
     vUTCOffset,
 )
 from icalendar.timezone import TZP, tzp
-from icalendar.tools import is_date
+from icalendar.tools import is_date, to_datetime
 
 if TYPE_CHECKING:
     from icalendar.alarms import Alarms
@@ -1232,6 +1232,8 @@ class Timezone(Component):
     way in which a time zone changes its offset from UTC over time.
     """
 
+    subcomponents: list[TimezoneStandard|TimezoneDaylight]
+
     name = "VTIMEZONE"
     canonical_order = ("TZID",)
     required = ("TZID",)  # it also requires one of components DAYLIGHT and STANDARD
@@ -1359,6 +1361,8 @@ class Timezone(Component):
         for component in self.walk():
             if type(component) == Timezone:
                 continue
+            if is_date(component["DTSTART"].dt):
+                component.DTSTART = to_datetime(component["DTSTART"].dt)
             assert isinstance(
                 component["DTSTART"].dt, datetime
             ), "VTIMEZONEs sub-components' DTSTART must be of type datetime, not date"